+++ ModelClient/Data/CategoryClient.cs
@@ -11,13 +11,13 @@ namespace ModelClient.Data
     public class CategoryClient : CategoryGlobal
     {
         public CategoryClient(){}
-        public CategoryClient(int id, TypeCat typeCat, float costPerDay)
+        public CategoryClient(int id, TypeCat typeCat, double costPerDay)
         {
             Id = id;
             TypeCat = typeCat;
             CostPerDay = costPerDay;
         }
-        public CategoryClient(TypeCat typeCat, float costPerDay)
+        public CategoryClient(TypeCat typeCat, double costPerDay)
         {
             TypeCat = typeCat;
             CostPerDay = costPerDay;
