+++ ModelClient/Data/ContractClient.cs
@@ -10,7 +10,7 @@ namespace ModelClient.Data
     public class ContractClient : ContractGlobal
     {
         public ContractClient(){}
-        public ContractClient(int id, int contractNum, DateTime realReturnDate, string departKm, string backKm, DateTime contractDate, float amountTotHTVA, float amountTotTVA, bool signed, int reservationId, int vehicleId, int penalizationId)
+        public ContractClient(int id, int contractNum, DateTime realReturnDate, string departKm, string backKm, DateTime contractDate, double amountTotHTVA, double amountTotTVA, bool signed, int reservationId, int vehicleId, int penalizationId)
         {
             Id = id;
             ContractNum = contractNum;
